requirements.txt
PyMuPDF
pymupdf4llm
PyPDF2
pdfplumber
pypdf

python-dotenv
pgvector
psycopg2
sqlalchemy
boto3